Fix serial isinstance() check for Python 3.9
We did isinstance(data, rich.console.RenderableType), with that being defined in
rich as: Union[ConsoleRenderable, RichCast, str]

Using Union with isinstance() is only possible starting in Python 3.10:
https://docs.python.org/3/whatsnew/3.10.html#pep-604-new-type-union-operator
https://peps.python.org/pep-0604/#isinstance-and-issubclass

In Python 3.9, this failed with:

    TypeError: Subscripted generics cannot be used with class and instance checks

Instead, we check only for rich.console.ConsoleRenderable (Protocol having a
__rich_console__) instead: We do already handle "str" earlier in the "if" (and
they should *not* be included here!); and we don't care about RichCast (Protocol
having a __rich__). All types we return as rich objects (rich.pretty.Pretty,
rich.table.Table, rich.console.Group) seem to have a __rich_console__ and thus
be ConsoleRenderables as well.

Thanks to Xarka on the PiShock Discord for catching this!